+++ zooFrontend/src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import App from './App';
 //import reportWebVitals from './reportWebVitals';
 import { BrowserRouter } from 'react-router-dom';
-import './index.css';
+import './styles/index.css';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
