+++ zooFrontend/src/pages/Caregiver/Dashboard/CaregiverDashboard.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import '../Dashboard/styles/Dashboards.css';
+import '../../../styles/Dashboards.css';
 
 const CaregiverDashboard = () => {
   return (
