+++ zooFrontend/src/pages/Director/Dashboard/DirectorDashboard.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import '../Dashboard/styles/Dashboards.css';
+import '../../../styles/Dashboards.css';
 
 const DirectorDashboard = () => {
   return (
