+++ zooFrontend/src/styles/Home.css
